requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
beautifulsoup4
playwright
pydantic
pydantic-settings
httpx
redis
celery
flower
python-dotenv
cryptography
pytest
pytest-asyncio
aiohttp
tenacity
loguru
pandas
numpy
openpyxl
sentry-sdk